# api/openapi.json
{"openapi":"3.1.0","info":{"title":"AI Teacher Engine API","description":"Cognitive Socratic Teaching Engine with RAG, Learner Model, and Misconception Adaptation.","version":"1.0.0"},"paths":{"/api/v1/health":{"get":{"tags":["System"],"summary":"Health Check","operationId":"health_check_api_v1_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/documents/upload":{"post":{"tags":["Knowledge Engine"],"summary":"Upload Document","description":"Parses, extracts sections, and chunks an uploaded document.","operationId":"upload_document_api_v1_documents_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_document_api_v1_documents_upload_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/documents":{"get":{"tags":["Knowledge Engine"],"summary":"List Documents","description":"Lists all documents available in the student's material library.","operationId":"list_documents_api_v1_documents_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/concepts/graph":{"get":{"tags":["Concept Graph"],"summary":"Get Concept Graph","description":"Returns the directed concept graph and prerequisite relations.","operationId":"get_concept_graph_api_v1_concepts_graph_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/students/knowledge-state":{"get":{"tags":["Learner Model"],"summary":"Get Student Knowledge State","description":"Returns current mastery percentages, confidence, and detected misconceptions.","operationId":"get_student_knowledge_state_api_v1_students_knowledge_state_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/lessons/create":{"post":{"tags":["Lesson Planner"],"summary":"Create Lesson","description":"Generates a personalized, structured multi-segment lesson plan.","operationId":"create_lesson_api_v1_lessons_create_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateLessonPayload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/teaching/checkpoint/answer":{"post":{"tags":["Teaching Engine"],"summary":"Submit Checkpoint Answer","description":"Evaluates student answer, updates mastery or triggers adaptive misconception state.","operationId":"submit_checkpoint_answer_api_v1_teaching_checkpoint_answer_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CheckpointAnswerPayload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/teaching/brain-inspect":{"get":{"tags":["Teacher Brain"],"summary":"Inspect Teacher Brain","description":"Returns real-time transparent educator parameters for judges/developers.","operationId":"inspect_teacher_brain_api_v1_teaching_brain_inspect_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/teaching/contextual-ask":{"post":{"tags":["Teaching Engine"],"summary":"Contextual Ask","description":"Contextual Q&A grounded directly in the current lesson concept.","operationId":"contextual_ask_api_v1_teaching_contextual_ask_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ContextualAskPayload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/visuals/render":{"post":{"tags":["Visual Engine"],"summary":"Render Visual","description":"Renders dynamic SVG visuals (electrical circuit or hydraulic water pipe).","operationId":"render_visual_api_v1_visuals_render_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RenderVisualPayload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/teaching/language/switch":{"post":{"tags":["Multilingual Engine"],"summary":"Switch Language","description":"Translates concept explanation and captions into English, Hindi, or Hinglish.","operationId":"switch_language_api_v1_teaching_language_switch_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SwitchLanguagePayload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/media/voice/synthesize":{"post":{"tags":["Media Engine"],"summary":"Synthesize Voice","description":"Synthesizes voice segment with status tracking and playback URL.","operationId":"synthesize_voice_api_v1_media_voice_synthesize_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SynthesizeVoicePayload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/media/fallback":{"post":{"tags":["Media Engine"],"summary":"Trigger Media Fallback","description":"Gracefully downgrades from full avatar video to audio + visual canvas + captions.","operationId":"trigger_media_fallback_api_v1_media_fallback_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MediaFallbackPayload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/assessment/generate":{"get":{"tags":["Assessment"],"summary":"Generate Assessment","description":"Generates an 8-question active diagnostic assessment based on concept graph.","operationId":"generate_assessment_api_v1_assessment_generate_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/assessment/submit":{"post":{"tags":["Assessment"],"summary":"Submit Assessment","description":"Evaluates student assessment and generates comprehensive editorial report.","operationId":"submit_assessment_api_v1_assessment_submit_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SubmitAssessmentPayload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/advanced/flashcards":{"get":{"tags":["Advanced Features"],"summary":"Get Flashcards","description":"Generates tactile Socratic flashcards for active recall revision.","operationId":"get_flashcards_api_v1_advanced_flashcards_get","parameters":[{"name":"topic","in":"query","required":false,"schema":{"type":"string","default":"Electricity","title":"Topic"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/advanced/notes":{"get":{"tags":["Advanced Features"],"summary":"Get Study Notes","description":"Generates Cornell-style study notes for the Warm Study Desk.","operationId":"get_study_notes_api_v1_advanced_notes_get","parameters":[{"name":"topic","in":"query","required":false,"schema":{"type":"string","default":"Electricity","title":"Topic"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/advanced/homework":{"post":{"tags":["Advanced Features"],"summary":"Generate Homework","description":"Generates targeted practice problems addressing diagnosed misconceptions.","operationId":"generate_homework_api_v1_advanced_homework_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GenerateHomeworkPayload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/advanced/revision":{"post":{"tags":["Advanced Features"],"summary":"Start Revision Mode","description":"Initializes a focused 5-minute targeted revision sprint with water-pipe tactile demonstrations.","operationId":"start_revision_mode_api_v1_advanced_revision_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StartRevisionPayload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/analytics/log":{"post":{"tags":["Production Support"],"summary":"Log Session Event","description":"Logs session telemetry and pedagogical events.","operationId":"log_session_event_api_v1_analytics_log_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LogAnalyticsPayload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/analytics/summary":{"get":{"tags":["Production Support"],"summary":"Get Analytics Summary","description":"Returns aggregated metrics and event counts for the educator dashboard.","operationId":"get_analytics_summary_api_v1_analytics_summary_get","parameters":[{"name":"session_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Session Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/media/cancel":{"post":{"tags":["Production Support"],"summary":"Cancel Media Job","description":"Cleanly terminates an in-flight video/voice generation task.","operationId":"cancel_media_job_api_v1_media_cancel_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobControlPayload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/media/retry":{"post":{"tags":["Production Support"],"summary":"Retry Media Job","description":"Retries a failed generation pipeline step.","operationId":"retry_media_job_api_v1_media_retry_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/JobControlPayload"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/accessibility/mode":{"post":{"tags":["Production Support"],"summary":"Set Accessibility Mode","description":"Persists student accessibility preferences (reduced motion, captions, contrast).","operationId":"set_accessibility_mode_api_v1_accessibility_mode_post","requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AccessibilityPayload"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"get":{"tags":["Production Support"],"summary":"Get Accessibility Mode","description":"Fetches active accessibility preferences.","operationId":"get_accessibility_mode_api_v1_accessibility_mode_get","parameters":[{"name":"student_id","in":"query","required":false,"schema":{"type":"string","default":"default_student","title":"Student Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"tags":["System"],"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"AccessibilityPayload":{"properties":{"student_id":{"type":"string","title":"Student Id","description":"Student ID","default":"default_student"},"settings":{"additionalProperties":true,"type":"object","title":"Settings","description":"Accessibility settings dictionary"}},"type":"object","required":["settings"],"title":"AccessibilityPayload"},"Body_upload_document_api_v1_documents_upload_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_document_api_v1_documents_upload_post"},"CheckpointAnswerPayload":{"properties":{"choice":{"type":"string","title":"Choice","description":"Student choice ('A', 'B', 'C', or 'D')"},"concept":{"type":"string","title":"Concept","description":"Active learning concept","default":"Resistance"}},"type":"object","required":["choice"],"title":"CheckpointAnswerPayload"},"ContextualAskPayload":{"properties":{"question":{"type":"string","title":"Question","description":"Student inquiry text"},"concept":{"type":"string","title":"Concept","description":"Active learning concept","default":"Resistance"}},"type":"object","required":["question"],"title":"ContextualAskPayload"},"CreateLessonPayload":{"properties":{"topic":{"type":"string","title":"Topic"},"level":{"type":"string","title":"Level","default":"beginner"},"duration_minutes":{"type":"integer","title":"Duration Minutes","default":20},"language":{"type":"string","title":"Language","default":"Hinglish"},"objective":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Objective"}},"type":"object","required":["topic"],"title":"CreateLessonPayload"},"GenerateHomeworkPayload":{"properties":{"weak_areas":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Weak Areas","description":"List of diagnosed misconception areas"}},"type":"object","title":"GenerateHomeworkPayload"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"JobControlPayload":{"properties":{"job_id":{"type":"string","title":"Job Id","description":"Target pipeline job identifier"}},"type":"object","required":["job_id"],"title":"JobControlPayload"},"LogAnalyticsPayload":{"properties":{"session_id":{"type":"string","title":"Session Id","description":"Active learning session ID","default":"demo_session"},"event_type":{"type":"string","title":"Event Type","description":"Action type ('pause', 'checkpoint_submit', etc.)"},"metadata":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Metadata"}},"type":"object","required":["event_type"],"title":"LogAnalyticsPayload"},"MediaFallbackPayload":{"properties":{"job_id":{"type":"string","title":"Job Id","description":"Target job identifier"},"error_reason":{"type":"string","title":"Error Reason","description":"Degradation cause","default":"Avatar video pipeline timeout"}},"type":"object","required":["job_id"],"title":"MediaFallbackPayload"},"RenderVisualPayload":{"properties":{"visual_type":{"type":"string","title":"Visual Type","default":"circuit"},"voltage":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Voltage","default":9.0},"resistance":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Resistance","default":10.0},"switch_closed":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Switch Closed","default":true},"pipe_width":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Pipe Width","default":"Narrow"}},"type":"object","title":"RenderVisualPayload"},"StartRevisionPayload":{"properties":{"topic":{"type":"string","title":"Topic","description":"Topic to revise","default":"Ohm's Law"}},"type":"object","title":"StartRevisionPayload"},"SubmitAssessmentPayload":{"properties":{"student_answers":{"anyOf":[{"additionalProperties":{"type":"string"},"type":"object"},{"type":"null"}],"title":"Student Answers","description":"Map of question_id -> selected_option (e.g. {'q1': 'B'})"}},"type":"object","title":"SubmitAssessmentPayload"},"SwitchLanguagePayload":{"properties":{"concept":{"type":"string","title":"Concept","description":"Concept name to translate","default":"Voltage"},"target_language":{"type":"string","title":"Target Language","description":"Target language ('English', 'Hindi', 'Hinglish')","default":"Hinglish"}},"type":"object","title":"SwitchLanguagePayload"},"SynthesizeVoicePayload":{"properties":{"text":{"type":"string","title":"Text","description":"Script text to synthesize"},"language":{"type":"string","title":"Language","description":"Spoken language","default":"Hinglish"},"voice_id":{"type":"string","title":"Voice Id","description":"Voice profile identifier","default":"dr_aris_calm"}},"type":"object","required":["text"],"title":"SynthesizeVoicePayload"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}