# backend/app/api/v1/endpoints/teaching.py
"""
Teaching Engine, Checkpoints & Teacher Brain Endpoints
"""

from fastapi import APIRouter, Depends
from app.core.deps import get_teaching_machine, get_learner_service
from app.teaching.state_machine import TeachingStateMachine
from app.services.learner_model import LearnerModelService
from app.schemas.teaching import CheckpointAnswerPayload, ContextualAskPayload

router = APIRouter()

@router.post("/teaching/checkpoint/answer", tags=["Teaching Engine"])
async def submit_checkpoint_answer(
    payload: CheckpointAnswerPayload,
    teaching_machine: TeachingStateMachine = Depends(get_teaching_machine),
    learner_service: LearnerModelService = Depends(get_learner_service)
):
    """
    Evaluates student answer, updates mastery or triggers adaptive misconception state.
    """
    result = teaching_machine.evaluate_checkpoint(payload.choice, concept=payload.concept)
    learner_service.update_mastery("resistance", is_correct=result["is_correct"])
    return result

@router.get("/teaching/brain-inspect", tags=["Teacher Brain"])
async def inspect_teacher_brain(
    teaching_machine: TeachingStateMachine = Depends(get_teaching_machine)
):
    """
    Returns real-time transparent educator parameters for judges/developers.
    """
    return teaching_machine.get_teacher_brain()

@router.post("/teaching/contextual-ask", tags=["Teaching Engine"])
async def contextual_ask(payload: ContextualAskPayload):
    """
    Contextual Q&A grounded directly in the current lesson concept.
    """
    return {
        "context": f"{payload.concept} · Module 03",
        "teacher_response": "Think of resistance as narrowing a water pipe: with the same water pressure, fewer gallons pass per minute. That is why current must drop when resistance increases.",
        "suggested_actions": ["Explain differently", "Show an example", "Give me a question"]
    }